requirements.txt
tendo==0.3.0
pyzmq==25.0.2
plugin==0.0.1
